refactor: move store into shared services
